requirements.txt
langchain>=0.2
langchain-community
langchain-openai
langchain-qdrant
qdrant-client
pypdf
python-dotenv
tiktoken
